Major additions. Check Desc
Additions:
Multiple tables for imported and queries
Multiple bugfixes
Added Print Form
Program Imports to database and sorts data

Current know bugs:
Options Screen:
Cannot Create Accounts
Cannot Edit Accounts
Cannot Delete Accounts
Cannot View Accounts

___________________________________________________
Current functions:
When other forms are shown the parent form should close.

Menu Form:
Open Import form in the position of the Menu form
Open Print form in the position of the Menu form. If the user hasn't
been on the Import Form in that session it should display a confirmation
dialog box first.

Open Login form

___________________________________________________
Login Form:
If fields are blank, relevant message box is shown
If Username is incorrect, relevant message box is show
If Password is incorrect when Username is correct, relevant message box
is shown

If Submit is clicked when Username and Password are correct then the
menu form should be shown if it is the initial log in else it should
show the options menu

If Submit is clicked it should clear the password field

If Cancel or X button is clicked on program start up program closes.

If Cancel or X button is clicked when transferring to Options form menu
form is shown

If Enter is pressed when on Username field it should jump to Password
field

If Enter is pressed when on Password field it should submit the username
and password
___________________________________________________
Import Form:
If Select File button is clicked then file dialog should show
If file is selected in file dialog then it should be imported; Putting
all of the data into the Imported table

If Import button is clicked then the program should make, in the
database, a table called simpleTable, containing simple orders, a table
called complexTable, containing complex orders, and a table called
pImport, containing all of the imported data. It should also update the
Simple and Complex tables to display the queried data.

If Print button is clicked the Print form should be shown
If Close or X button is clicked the Menu form should be shown

___________________________________________________
Print Form:
When form is loaded the fields should display the information from the
first record, the Record label should display the current record and the
Records label should display the total number of records.

If the next button is clicked, the next record should be shown and the
Record label should update accordingly. If the record is at the last
record then it should show the first record.

If the prev button is clicked, the previous record should be shown and
the Record label should update accordingly. If the record is at the
first record then it should show the last record.

If X button is clicked the Menu form should be shown